add the first version of the ast
